refactor: address reviewer comments for Mindray HL7 PR
1. AnalyzerImportController: Improve error handling
   - Set content type before writing response
   - Use sendError() for error cases per servlet best practices
   - Keep defensive null check for reader

2. Tests: Extract common logic to AbstractMindrayIntegrationTest
   - Reduce duplication between BC-5380 and BS-360E tests
   - Common setup, cleanup, and verification in base class
   - Subclasses only specify analyzer IP and fixture path
